전체활동 리스트 세팅 및 인증시 데이터 올리기

+++ app/src/main/java/com/swu/dimiz/ogg/oggdata/remotedatabase/MyAllAct.kt
@@ -12,5 +12,5 @@ data class MyAllAct ( //전체활동 활동명 별 상황 저장 -> graph에서 
 
     var upCount:Int =0,
 
-    var allC02:Float =0f
+    var allC02:Double =0.0
 )
